tests(iam_application): enable sweeper and update tests

+++ scaleway/resource_iam_application_test.go
@@ -11,9 +11,6 @@ import (
 )
 
 func init() {
-	if !terraformBetaEnabled {
-		return
-	}
 	resource.AddTestSweepers("scaleway_iam_application", &resource.Sweeper{
 		Name: "scaleway_iam_application",
 		F:    testSweepIamApplication,
@@ -29,6 +26,10 @@ func testSweepIamApplication(_ string) error {
 			return fmt.Errorf("failed to list applications: %w", err)
 		}
 		for _, app := range listApps.Applications {
+			if !isTestResource(app.Name) {
+				continue
+			}
+
 			err = api.DeleteApplication(&iam.DeleteApplicationRequest{
 				ApplicationID: app.ID,
 			})
